% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Visualise.R
\name{PlotDensities.SplitScore_job}
\alias{PlotDensities.SplitScore_job}
\title{Visualise channel density estimates for SplitScore metaclusters}
\usage{
\method{PlotDensities}{SplitScore_job}(ssj, cols = NULL)
}
\arguments{
\item{ssj}{\code{SplitScore_job} object}

\item{cols}{optional character vector of column names or numeric vector of column indices specifying parameters of input expression data for which to plot density estimates. Default value is \code{NULL}, whereby plots are generated for all parameters}
}
\description{
Plots a grid of density plots for select (or all) marker expression (channel signals) for SplitScore-generated metaclusters.
}
