% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Prune.R
\name{Prune}
\alias{Prune}
\title{Perform metaclustering by pruning a hierarchical clustering dendrogram}
\usage{
Prune(ssj, fun.split_score, n_metaclusters, verbose = 1L, seed = NULL)
}
\arguments{
\item{ssj}{\code{SplitScore_job} object}

\item{fun.split_score}{function that takes indices of a subtree and its two subtrees and returns a numeric score}

\item{n_metaclusters}{integer value (\code{>2}): target number of generated metaclusters}

\item{verbose}{integer value: level of verbosity (currently either \code{0} or a positive value). Default value is \code{1}}

\item{seed}{optional numeric value: random seed. Default value is \code{NULL}}
}
\description{
Prunes the dendrogram created by applying \code{AgglomerativeClustering} to a \code{SplitScore_job} object.
}
